@@ -1,20 +1,21 @@
 name: dropthenumber
 description: A game to play 2048 vertically.
 publish_to: "none"
-version: 1.0.1+2
+version: 1.0.2
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.3.0
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  flame: ^0.29.4
-  flutter_screenutil: ^5.0.0+2
-  sprintf: "^4.0.0"
-  audioplayers: ^0.18.3
-  shared_preferences: ^2.0.8
+  cupertino_icons: ^1.0.6
+  flame: ^1.16.0
+  flutter_screenutil: ^5.9.0
+  sprintf: ^7.0.0
+  audioplayers: ^5.2.1
+  shared_preferences: ^2.2.2
+  flame_audio: ^2.10.0
 
 dev_dependencies:
   flutter_test:
