@@ -1,6 +1,5 @@
-// @dart=2.11
 import 'package:flame/flame.dart';
-import 'package:flame/util.dart';
+import 'package:flame/game.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -8,10 +7,9 @@ import 'dropthenumber.dart';
 import 'data_handler.dart';
 
 void main() async {
-  Util flameUtil = Util();
-  flameUtil.fullScreen();
-  flameUtil.setOrientation(DeviceOrientation.portraitUp);
-  flameUtil.setPortraitUpOnly();
+  Flame.device.fullScreen();
+  Flame.device.setOrientation(DeviceOrientation.portraitUp);
+  Flame.device.setPortraitUpOnly();
 
   // SharedPreferences "Null check operator used on a null value" problem
   // Solution found: https://stackoverflow.com/a/62493934
@@ -19,5 +17,5 @@ void main() async {
 
   SharedPreferences storage = await SharedPreferences.getInstance();
   DropTheNumber dropTheNumber = DropTheNumber(DataHandler(storage));
-  runApp(dropTheNumber.widget);
+  runApp(GameWidget(game: dropTheNumber));
 }
